refactor: move nodes logic to src
